Add 'networking' dir, move files into it.
