+++ frontend/lib/forms/geo-autocomplete.tsx
@@ -5,7 +5,7 @@ import autobind from 'autobind-decorator';
 import { BoroughChoice, getBoroughChoiceLabels } from '../../../common-data/borough-choices';
 import { WithFormFieldErrors, formatErrors } from './form-errors';
 import { bulmaClasses } from '../ui/bulma';
-import { awesomeFetch, createAbortController } from '../fetch';
+import { awesomeFetch, createAbortController } from '../networking/fetch';
 import { renderLabel, LabelRenderer } from './form-fields';
 import { KEY_ENTER, KEY_TAB } from '../key-codes';
 import { GeoSearchBoroughGid, GeoSearchResults, GeoSearchRequester } from '@justfixnyc/geosearch-requester';
