+++ frontend/lib/networking/custom-graphql-scalars.d.ts
