+++ frontend/lib/networking/query-loader-prefetcher.tsx
@@ -1,8 +1,8 @@
 import { GraphQLFetch } from "./graphql-client";
 import { RouteComponentProps } from "react-router";
-import { AppContextType } from "./app-context";
-import { isDeepEqual } from "./util";
-import { getAppStaticContext } from "./app-static-context";
+import { AppContextType } from "../app-context";
+import { isDeepEqual } from "../util";
+import { getAppStaticContext } from "../app-static-context";
 
 export interface QueryLoaderFetch<Input, Output> {
   (fetch: GraphQLFetch, args: Input): Promise<Output>;
