+++ frontend/lib/networking/session-poller.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 
-import { AllSessionInfo } from "./queries/AllSessionInfo";
+import { AllSessionInfo } from "../queries/AllSessionInfo";
 import { GraphQLFetch } from "./graphql-client";
-import { AppContextType, withAppContext } from './app-context';
+import { AppContextType, withAppContext } from '../app-context';
 import autobind from 'autobind-decorator';
 import { Helmet } from 'react-helmet-async';
-import { NoScriptFallback } from './progressive-enhancement';
+import { NoScriptFallback } from '../progressive-enhancement';
 
 const DEFAULT_INTERVAL_MS = 5000;
