+++ frontend/lib/pages/verify-email.tsx
@@ -7,7 +7,7 @@ import { AppContext } from '../app-context';
 import Page from '../ui/page';
 import { RouteComponentProps, withRouter, Redirect, Link } from 'react-router-dom';
 import { getQuerystringVar } from '../querystring';
-import { SessionPoller, SessionPollerProps } from '../session-poller';
+import { SessionPoller, SessionPollerProps } from '../networking/session-poller';
 import { GetEmailVerificationStatus } from '../queries/GetEmailVerificationStatus';
 import { MiddleProgressStep } from '../progress/progress-step-route';
 import { SimpleProgressiveEnhancement } from '../progressive-enhancement';
